requirements.txt
appnope==0.1.0
arrow==0.15.5
backports.shutil-get-terminal-size==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
decorator==4.4.2
Django==1.11.29
django-braces==1.13.0
django-crispy-forms==1.7.2
django-floppyforms==1.7.0
django-ranged-response==0.2.0
django-recaptcha==2.0.4
django-reversion==3.0.4
django-tinymce4-lite==1.7.1
django-vanilla-views==1.0.5
djangorestframework==3.9.1
djangorestframework-jsonp==1.0.2
enum34==1.1.10
geopy==1.21.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jsmin==2.2.2
lxml==4.5.0
pathlib2==2.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
Pillow-PIL==0.1.dev0
prompt-toolkit==3.0.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.5.2
pychrome==0.2.3
Pygments==2.6.1
pyquery==1.4.1
python-dateutil==2.8.1
python-subprocess2==2.0.2
pytz==2019.3
requests==2.23.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.14.0
South==1.0.2
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.9
websocket-client==0.57.0
xlrd==1.2.0
xlwt==1.3.0
python-quickbooks==0.7.5
beautifulsoup4==4.9.0
html-table-extractor==1.4.0
sentry-sdk==0.14.3
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4